+++ app/src/test/java/com/jaac/avoqado_tpv/features/payment/presentation/PaymentViewModelTest.kt
@@ -309,6 +309,7 @@ class PaymentViewModelTest {
             modulesRepository = mockModulesRepository,
             connectionStateManager = mockConnectionStateManager,
             merchantRepository = mockk(relaxed = true),
+            merchantEligibilityRepository = mockk(relaxed = true),
             criticalNetworkOperationManager = mockCriticalNetworkOperationManager,
             connectionEventManager = mockConnectionEventManager,
             appContext = mockAppContext