+++ app/src/test/java/com/jaac/avoqado_tpv/features/payment/presentation/angelpay/AngelPayPaymentViewModelTest.kt
@@ -170,6 +170,7 @@ class AngelPayPaymentViewModelTest {
         shiftRepository = shiftRepository,
         authRepository = authRepository,
         merchantRepository = merchantRepository,
+        merchantEligibilityRepository = mockk(relaxed = true),
         secureStorage = secureStorage,
         terminalConfigRepository = terminalConfigRepository,
         intentBuilder = intentBuilder,